pyproject.toml
[project]
name = "reading-flo-story-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.5.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
    "google-cloud-texttospeech>=2.16.0",
    "pydub>=0.25.1",
]
